agent/requirements.txt
slack-bolt>=1.18.0
slack-sdk>=3.27.0
notion-client>=2.2.0
python-dotenv>=1.0.0
